photo-book-generator/requirements.txt
# Core Dependencies
pillow>=10.0.0
exifread>=3.0.0
piexif>=1.1.3

# Computer Vision & Deep Learning
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
open-clip-torch>=2.20.0

# Clustering & Similarity
numpy>=1.24.0
scikit-learn>=1.3.0


# Graph & Storyline Logic
networkx>=3.1

# PDF Rendering
reportlab>=4.0.0

# Utilities
pydantic>=2.0.0
typer>=0.9.0
tqdm>=4.65.0
